# add basic rle encoding to make data transmit faster