--- a/Cargo.toml
+++ b/Cargo.toml
@@ -13,7 +13,7 @@ bench = false
 doc = false
 
 [dependencies]
-typst = { git = "https://github.com/typst/typst.git", tag = "v0.2.0"}
+typst = { git = "https://github.com/typst/typst.git", tag = "v0.2.0" }
 typst-library = { git = "https://github.com/typst/typst.git", tag = "v0.2.0" }
 chrono = { version = "0.4", default-features = false, features = [
   "clock",
@@ -39,6 +39,9 @@ log = "0.4.17"
 serde_json = "1.0.95"
 futures = "0.3.28"
 serde = "1.0.159"
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["local-time", "time"] }
+tracing-chrome = "0.7.1"
 
 [build-dependencies]
 clap = { version = "4.2.1", features = ["derive", "string"] }