before: 
- rust side: ~100ms render, ~200ms transmit
- browser side: ?ms render

after:
- rust side: ~100ms render, ~50ms encode & transmit
- browser side: ~100ms decode + render