@rapiz1 2023-04-27
200ms to transfer a framebuffer on local loopback? The number does not make sense.