@Enter-tainer 2023-04-27
Closed. This is no longer needed because we only send needed frames now(rather than all of them). The performance data is outdated either.